Introduced new feature: custom controller model Use the NTM-Layer with your own controller! How cool is that!
Well, you can take your own model of appropriate size and use it as an controller.
Only stateless controllers are supported right now, but recurrent controllers should be fine, as long as they are "statefull", which means they carry around their own state over batches. But how does statefulness work on model-level (on recurrent-layer-level it is easy)?
We'll have to see.